adding switch mode to set_locations, plus tests
